Refactor query and mutation schema / resolvers
